haihe-water-platform/backend/requirements.txt
Flask==2.0.3
Flask-CORS==3.0.10
Werkzeug==2.0.3
pymysql==1.0.2
neo4j==5.19.0
pyjwt==2.4.0
bcrypt==4.0.1
python-dotenv==0.20.0
requests==2.28.2
beautifulsoup4==4.11.1
lxml==4.9.1